+++ Rubrik/Public/Get-RubrikLogShipping.ps1
@@ -31,6 +31,11 @@ function Get-RubrikLogShipping
       Get-RubrkLogShipping -location am1-chrilumn-w1.rubrikdemo.com\MSSQLSERVER
 
       Get all log shipping configurations for a given location (log shipping secondary server)
+
+      .EXAMPLE
+      Get-RubrkLogShipping -location am1-chrilumn-w1.rubrikdemo.com\MSSQLSERVER -DetailedObject
+
+      Get all log shipping configurations for a given location (log shipping secondary server) and retrieves detailed information about these objects
   #>
 
   [CmdletBinding()]
@@ -62,6 +67,9 @@ function Get-RubrikLogShipping
     
     [ValidateSet("asc", "desc")]
     [String]$sort_order,
+
+    # DetailedObject will retrieved the detailed LogShipping object, the default behavior of the API is to only retrieve a subset of the full LogShipping object unless we query directly by ID. Using this parameter does affect performance as more data will be retrieved and more API-queries will be performed.
+    [Switch]$DetailedObject,
 
     # Rubrik server IP or FQDN
     [String]$Server = $global:RubrikConnection.server,
@@ -100,7 +108,16 @@ function Get-RubrikLogShipping
     $result = Test-FilterObject -filter ($resources.Filter) -result $result
     $result = Set-ObjectTypeName -TypeName $resources.ObjectTName -result $result
 
-    return $result
+    # If the Get-RubrikLogShipping function has been called with the -DetailedObject parameter a separate API query will be performed if the initial query was not based on ID
+    if (($DetailedObject) -and (-not $PSBoundParameters.containskey('id'))) {
+      for ($i = 0; $i -lt @($result).Count; $i++) {
+        $Percentage = [int]($i/@($result).count*100)
+        Write-Progress -Activity "DetailedObject queries in Progress, $($i+1) out of $(@($result).count)" -Status "$Percentage% Complete:" -PercentComplete $Percentage
+        Get-RubrikLogShipping -id $result[$i].id
+      }
+    } else {
+      return $result
+    }
 
   } # End of process
 } # End of function